+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static lotto.LottoNumber.MAX_LOTTO_NUMBER;
-import static lotto.LottoNumber.MIN_LOTTO_NUMBER;
-import static lotto.LottoNumbers.LOTTO_SIZE;
+import static lotto.domain.LottoNumber.MAX_LOTTO_NUMBER;
+import static lotto.domain.LottoNumber.MIN_LOTTO_NUMBER;
+import static lotto.domain.LottoNumbers.LOTTO_SIZE;
 
 public class LottoMachine {
     private static final List<LottoNumber> LOTTO_NUMBERS = IntStream.range(MIN_LOTTO_NUMBER, MAX_LOTTO_NUMBER + 1)
