+++ src/main/java/lotto/domain/LottoStore.java
@@ -1,8 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 public class LottoStore {
     public static final Integer LOTTO_PRICE = 1000;
