+++ src/main/java/lotto/view/InputView.java
@@ -6,8 +6,8 @@ import java.util.Scanner;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static lotto.LottoNumbers.LOTTO_SIZE;
-import static lotto.LottoStore.LOTTO_PRICE;
+import static lotto.domain.LottoNumbers.LOTTO_SIZE;
+import static lotto.domain.LottoStore.LOTTO_PRICE;
 
 public class InputView {
     private static final Scanner scanner = new Scanner(System.in);
