+++ src/test/java/lotto/LottoStoreTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoStore;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
